add user model, currency attribute and wallet amount in the smallest unit add specific exceptions for validations

+++ src/main/java/com/dilshan/wallet/dto/WalletDto.java
@@ -11,5 +11,5 @@ public class WalletDto {
     private long walletId;
 
 
-    private double amount;
+    private int amount;
 }
